requirements.txt
requests
pytz
pystray
pillow